Improvements

+++ packages/isar_core/src/core/instance.rs
@@ -5,9 +5,9 @@ use super::query_builder::IsarQueryBuilder;
 use super::schema::IsarSchema;
 
 pub struct CompactCondition {
-    pub min_file_size: u64,
-    pub min_bytes: u64,
-    pub min_ratio: f64,
+    pub min_file_size: u32,
+    pub min_bytes: u32,
+    pub min_ratio: f32,
 }
 
 pub trait IsarInstance {
@@ -29,14 +29,14 @@ pub trait IsarInstance {
     where
         Self: 'a;
 
-    fn get(instance_id: u64) -> Option<Self::Instance>;
+    fn get(instance_id: u32) -> Option<Self::Instance>;
 
     fn open(
-        instance_id: u64,
+        instance_id: u32,
         name: &str,
         dir: &str,
         schema: IsarSchema,
-        max_size_mib: usize,
+        max_size_mib: u32,
         relaxed_durability: bool,
         compact_condition: Option<CompactCondition>,
     ) -> Result<Self::Instance>;
@@ -47,18 +47,14 @@ pub trait IsarInstance {
 
     fn abort_txn(&self, txn: Self::Txn);
 
-    fn insert(
-        &self,
-        txn: Self::Txn,
-        collection_index: usize,
-        count: usize,
-    ) -> Result<Self::Insert<'_>>;
+    fn insert(&self, txn: Self::Txn, collection_index: u16, count: u32)
+        -> Result<Self::Insert<'_>>;
 
-    fn build_query(&self, collection_index: usize) -> Result<Self::QueryBuilder<'_>>;
+    fn build_query(&self, collection_index: u16) -> Result<Self::QueryBuilder<'_>>;
 
     fn query<'a>(&'a self, txn: &'a Self::Txn, query: &'a Self::Query) -> Result<Self::Cursor<'_>>;
 
-    fn count(&self, txn: &Self::Txn, query: &Self::Query) -> Result<usize>;
+    fn count(&self, txn: &Self::Txn, query: &Self::Query) -> Result<u32>;
 
-    fn delete(&self, txn: &Self::Txn, query: &Self::Query) -> Result<usize>;
+    fn delete(&self, txn: &Self::Txn, query: &Self::Query) -> Result<u32>;
 }
