+++ packages/isar_core/src/core/query_builder.rs
@@ -11,11 +11,11 @@ pub trait IsarQueryBuilder {
 
     fn set_filter(&mut self, filter: Filter);
 
-    fn add_sort(&mut self, property_index: usize, sort: Sort);
+    fn add_sort(&mut self, property_index: u16, sort: Sort);
 
-    fn set_offset(&mut self, offset: usize);
+    fn set_offset(&mut self, offset: u32);
 
-    fn set_limit(&mut self, limit: usize);
+    fn set_limit(&mut self, limit: u32);
 
     fn build(self) -> Self::Query;
 }
