+++ packages/isar_core/src/filter/filter_nested.rs
@@ -2,12 +2,12 @@ use super::Filter;
 
 #[derive(PartialEq, Clone, Debug)]
 pub struct FilterNested {
-    collection_index: usize,
+    collection_index: u16,
     filter: Box<Filter>,
 }
 
 impl FilterNested {
-    pub fn new(collection_index: usize, filter: Filter) -> Self {
+    pub fn new(collection_index: u16, filter: Filter) -> Self {
         FilterNested {
             collection_index,
             filter: Box::new(filter),
